do not generate globals twice
