+++ src/generators/string.rs
@@ -5,7 +5,7 @@ use super::{Context, KeyGenerator, SchemaKey};
 pub fn key_generator(key: &SchemaKey, aux_visibility: syn::Visibility) -> KeyGenerator<'_> {
     if let Some(ref choices) = key.choices {
         let choice_enum_name = key.name.to_pascal_case();
-        let choice_enum_token_stream = super::new_variant_enum(
+        let choice_enum_token_stream = super::enum_token_stream(
             &choice_enum_name,
             &choices
                 .choices
